gpg: Replace use of PRIu64 in log_debug
* g10/cipher-aead.c (do_flush): Use %llu and a cast.
* g10/decrypt-data.c (aead_underflow): Ditto.
--

Fixes-commit: b2cedc108d5cabb07f496c31c11c9cba6f328f76

We don't use the system's printf but the one implemented by
us (gpgrt's estream-printf) thus the PRIu64 may or may not be correct.
We can't do much about the -Wformat errors due to our different
implementation.

+++ g10/cipher-aead.c
@@ -295,9 +295,9 @@ do_flush (cipher_filter_context_t *cfx, iobuf_t a, byte *buf, size_t size)
           size_t n1 = cfx->chunksize - (cfx->chunklen + cfx->buflen);
           finalize = 1;
           if (DBG_FILTER)
-            log_debug ("chunksize %"PRIu64" reached;"
+            log_debug ("chunksize %llu reached;"
                        " cur buflen=%zu using %zu of %zu\n",
-                       cfx->chunksize, cfx->buflen,
+                       (unsigned long long)cfx->chunksize, cfx->buflen,
                        n1, n);
           n = n1;
         }
